chore(Action): implement a component to manage the different action
