+++ packages/components/src/Actions/ActionButton/index.js
@@ -1,0 +1,3 @@
+import ActionButton from './ActionButton.component';
+
+export default ActionButton;
