+++ packages/components/src/Actions/Actions.component.js
@@ -3,13 +3,13 @@ import React from 'react';
 import { ButtonGroup } from 'react-bootstrap';
 import classNames from 'classnames';
 import Action from './Action';
-import ActionDropdown from './ActionDropdown';
-import ActionSplitDropdown from './ActionSplitDropdown';
+import ActionButton from './ActionButton';
 
+console.dir(Action);
+console.dir(ActionButton);
 
 const TYPE_DROPDOWN = 'dropdown';
 const TYPE_SPLIT_DROPDOWN = 'splitDropdown';
-
 
 function getButtonGroupProps(props) {
 	const buttonGroupProps = {};
@@ -19,17 +19,6 @@ function getButtonGroupProps(props) {
 		}
 	});
 	return buttonGroupProps;
-}
-
-function getActionComponent(displayMode) {
-	switch (displayMode) {
-		case TYPE_DROPDOWN:
-			return ActionDropdown;
-		case TYPE_SPLIT_DROPDOWN:
-			return ActionSplitDropdown;
-		default:
-			return Action;
-	}
 }
 
 /**
@@ -84,22 +73,17 @@ function Actions(props) {
 	const buttonGroupProps = getButtonGroupProps(props);
 
 	return (
-		<ButtonGroup
-			className={classNames('tc-actions', props.className)}
-			{...buttonGroupProps}
-		>
+		<ButtonGroup className={classNames('tc-actions', props.className)} {...buttonGroupProps}>
 			{props.actions.map((action, index) => {
-				const { displayMode, ...rest } = action;
-				const ActionComponent = getActionComponent(displayMode);
 				const params = {
 					key: index,
 					hideLabel: props.hideLabel,
 					link: props.link,
 					tooltipPlacement: props.tooltipPlacement,
-					...rest,
+					...action,
 				};
 
-				return <ActionComponent {...params} />;
+				return <Action {...params} />;
 			})}
 		</ButtonGroup>
 	);
@@ -108,7 +92,7 @@ function Actions(props) {
 Actions.propTypes = {
 	actions: PropTypes.arrayOf(
 		PropTypes.oneOfType([
-			PropTypes.shape(Action.propTypes),
+			PropTypes.shape(ActionButton.propTypes),
 			PropTypes.shape({
 				displayMode: PropTypes.oneOf([TYPE_DROPDOWN, TYPE_SPLIT_DROPDOWN]),
 			}),
@@ -116,7 +100,7 @@ Actions.propTypes = {
 	),
 	className: PropTypes.string,
 	hideLabel: PropTypes.bool,
-	tooltipPlacement: Action.propTypes.tooltipPlacement,
+	tooltipPlacement: ActionButton.propTypes.tooltipPlacement,
 	link: PropTypes.bool,
 	...ButtonGroup.propTypes,
 };
