+++ packages/components/src/index.js
@@ -68,7 +68,7 @@ import {
 
 import I18N_DOMAIN_COMPONENTS from './constants';
 
-import { Action, ActionDropdown, ActionSplitDropdown, Actions } from './Actions';
+import { Action, ActionDropdown, ActionButton, ActionSplitDropdown, Actions } from './Actions';
 import ActionBar from './ActionBar';
 import Notification from './Notification';
 import AppHeaderBar from './AppHeaderBar';
@@ -105,6 +105,7 @@ import WithDrawer from './WithDrawer';
 export {
 	Action,
 	ActionBar,
+	ActionButton,
 	ActionDropdown,
 	ActionSplitDropdown,
 	Actions,
@@ -139,7 +140,6 @@ export {
 	Typeahead,
 	VirtualizedList,
 	WithDrawer,
-
 	Alert,
 	Badge as BootstrapBadge,
 	Breadcrumb,
